1.支持配置跨工程替换（包含对加解密的处理） 2.支持非关键配置项隐藏 3.支持界面操作配置项更换模块 4.增加分页个数控制[5,10,15,20] 5.修改了配置导出到.properties中的格式以及代码逻辑 6.在新建项目时增加了是否为公共项目的选项以及在添加配置时增加了是否在列表中隐藏的选项 7.修改了建表脚本，在表conf_project中增加了IS_COMMON字段以及在conf_project_config表中增加了IS_SHOW字段 8.修改了版本号为1.3.0-SNAPSHOT

+++ super-diamond-server/src/main/java/com/github/diamond/web/model/ModuleIdExist.java
@@ -6,18 +6,18 @@ package com.github.diamond.web.model;
 public class ModuleIdExist {
 
     /**
-     * 判断moduleId是否存在
+     * 判断moduleId是否存在.
      */
 
     private boolean isExist;
 
 
     /**
-     * 模型Id
+     * 模型Id.
      */
-    private long moduleId;
+    private int moduleId;
 
-    public ModuleIdExist(boolean isExist, long moduleId) {
+    public ModuleIdExist(boolean isExist, int moduleId) {
         this.isExist = isExist;
         this.moduleId = moduleId;
     }
@@ -30,11 +30,11 @@ public class ModuleIdExist {
         this.isExist = isExist;
     }
 
-    public long getModuleId() {
+    public int getModuleId() {
         return moduleId;
     }
 
-    public void setModuleId(long moduleId) {
+    public void setModuleId(int moduleId) {
         this.moduleId = moduleId;
     }
 }
