+++ super-diamond-server/src/main/java/com/github/diamond/web/model/Project.java
@@ -3,60 +3,74 @@ package com.github.diamond.web.model;
 import java.util.List;
 
 /**
- * Create on @2013-7-18 @下午10:51:17 
+ * Create on @2013-7-18 @下午10:51:17.
+ *
  * @author melin
  */
 public class Project {
-	private long id;
-	private String code;
-	private String name;
-	private long ownerId;
-	
-	private String userCode;
-	private String userName;
-	
-	private List<String> roles;
-	
-	public long getId() {
-		return id;
-	}
-	public void setId(long id) {
-		this.id = id;
-	}
-	public String getCode() {
-		return code;
-	}
-	public void setCode(String code) {
-		this.code = code;
-	}
-	public String getName() {
-		return name;
-	}
-	public void setName(String name) {
-		this.name = name;
-	}
-	public long getOwnerId() {
-		return ownerId;
-	}
-	public void setOwnerId(long ownerId) {
-		this.ownerId = ownerId;
-	}
-	public String getUserCode() {
-		return userCode;
-	}
-	public void setUserCode(String userCode) {
-		this.userCode = userCode;
-	}
-	public String getUserName() {
-		return userName;
-	}
-	public void setUserName(String userName) {
-		this.userName = userName;
-	}
-	public List<String> getRoles() {
-		return roles;
-	}
-	public void setRoles(List<String> roles) {
-		this.roles = roles;
-	}
+    private int id;
+    private String code;
+    private String name;
+    private int ownerId;
+
+    private String userCode;
+    private String userName;
+
+    private List<String> roles;
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getCode() {
+        return code;
+    }
+
+    public void setCode(String code) {
+        this.code = code;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public int getOwnerId() {
+        return ownerId;
+    }
+
+    public void setOwnerId(int ownerId) {
+        this.ownerId = ownerId;
+    }
+
+    public String getUserCode() {
+        return userCode;
+    }
+
+    public void setUserCode(String userCode) {
+        this.userCode = userCode;
+    }
+
+    public String getUserName() {
+        return userName;
+    }
+
+    public void setUserName(String userName) {
+        this.userName = userName;
+    }
+
+    public List<String> getRoles() {
+        return roles;
+    }
+
+    public void setRoles(List<String> roles) {
+        this.roles = roles;
+    }
 }
